+++ packages/create-keystone-app/example-projects/starter/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/example-projects-starter
+
 # @keystone-alpha/example-projects-starter
 
 ## 1.1.15